move stories to stories/Checkbox path
