+++ packages/react-components/react-checkbox/src/stories/Checkbox/CheckboxCircular.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Checkbox } from './index';
+import { Checkbox } from '../../index';
 
 export const Circular = () => <Checkbox shape="circular" label="Circular" />;
 Circular.parameters = {
