+++ packages/react-components/react-checkbox/src/stories/Checkbox/CheckboxLabelBefore.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Checkbox } from './index';
+import { Checkbox } from '../../index';
 
 export const LabelBefore = () => <Checkbox labelPosition="before" label="Label before" />;
 LabelBefore.parameters = {
